# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Switch Drive Pairing API
  version: 1.0.0
servers:
  - url: https://api.erok.qzz.io
paths:
  /health:
    get:
      responses: {'200': {description: Healthy}}
  /v1/pairings:
    post:
      summary: Start a ten-minute Switch-to-phone pairing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [consolePublicKey]
              properties: {consolePublicKey: {type: string}}
      responses:
        '201':
          description: Pairing created
          content:
            application/json:
              schema:
                type: object
                required: [id, verificationUri, code, pollSecret, expiresAt]
                properties:
                  id: {type: string}
                  verificationUri: {type: string, format: uri}
                  code: {type: string}
                  pollSecret: {type: string}
                  expiresAt: {type: string, format: date-time}
  /v1/pairings/{id}:
    get:
      security: [{PairingSecret: []}]
      responses:
        '200':
          description: Pairing status or approved Google account
        '410': {description: Expired}
  /v1/pairings/{id}/claim:
    post:
      security: [{PairingSecret: []}]
      responses:
        '200': {description: Console session and approved Google account}
        '409': {description: Not approved or already consumed}
  /v1/accounts:
    get:
      security: [{ConsoleSession: []}]
      responses: {'200': {description: Linked Google accounts}}
  /v1/accounts/{id}/access-token:
    post:
      security: [{ConsoleSession: []}]
      responses: {'200': {description: Short-lived Drive access token}}
  /v1/accounts/{id}:
    delete:
      security: [{ConsoleSession: []}]
      responses: {'204': {description: Account unlinked; the final link also deletes stored authorization and attempts Google revocation}}
components:
  schemas:
    Account:
      type: object
      required: [id, email, displayName]
      properties:
        id: {type: string, format: uuid}
        email: {type: string, format: email}
        displayName: {type: [string, 'null']}
  securitySchemes:
    PairingSecret: {type: apiKey, in: header, name: X-Pairing-Secret}
    ConsoleSession: {type: http, scheme: bearer}